move callback error handler under Utility in Tsavorite
